LayoutTests/imported/w3c/web-platform-tests/mediacapture-fromelement/capture-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.

Harness Error (FAIL), message = Unhandled rejection: The operation is not supported.

PASS video-only
FAIL capture promise_test: Unhandled rejection with value: object "TypeError: video.captureStream is not a function. (In 'video.captureStream()', 'video.captureStream' is undefined)"
PASS audio-only
FAIL capture 1 promise_test: Unhandled rejection with value: object "TypeError: video.captureStream is not a function. (In 'video.captureStream()', 'video.captureStream' is undefined)"
PASS video+audio
FAIL capture 2 promise_test: Unhandled rejection with value: object "TypeError: video.captureStream is not a function. (In 'video.captureStream()', 'video.captureStream' is undefined)"
